Convert L64x0 to Zephyr device
This commit looks huge but it's actually a simpl conversion of L64x0
from an ad-hoc SPI device to a Zephyr device.  We can now do

    struct device *dev = DEVICE_DT_GET(MOTOR_DRIVER);

To get the device and call functions on it.

As a result, all SPI bits are encapslated in l64x0.c and device
tree (app.overlay in this case) and not leaked to the application
layer.

And finally we have copyright notices in the source files.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>